Make sure that all Anaconda's DBus errors are registered
Register all DBus errors right after the error mapper is created.
