+++ pyanaconda/modules/common/errors/general.py
@@ -1,7 +1,7 @@
 #
-# DBus errors related to tasks.
+# General DBus errors.
 #
-# Copyright (C) 2018  Red Hat, Inc.  All rights reserved.
+# Copyright (C) 2020  Red Hat, Inc.  All rights reserved.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,18 +16,25 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
+from dasbus.error import DBusError
+
 from pyanaconda.core.dbus import dbus_error
 from pyanaconda.modules.common.constants.namespaces import ANACONDA_NAMESPACE
-from pyanaconda.modules.common.errors.general import AnacondaError
 
 
-@dbus_error("TaskError", namespace=ANACONDA_NAMESPACE)
-class TaskError(AnacondaError):
-    """General exception for task errors."""
+@dbus_error("Error", namespace=ANACONDA_NAMESPACE)
+class AnacondaError(DBusError):
+    """A default Anaconda error."""
     pass
 
 
-@dbus_error("NoResultError", namespace=ANACONDA_NAMESPACE)
-class NoResultError(TaskError):
-    """There is no result of a task."""
+@dbus_error("InvalidValueError", namespace=ANACONDA_NAMESPACE)
+class InvalidValueError(AnacondaError):
+    """Invalid value passed."""
+    pass
+
+
+@dbus_error("UnsupportedValueError", namespace=ANACONDA_NAMESPACE)
+class UnsupportedValueError(AnacondaError):
+    """Value passed is not supported."""
     pass
