+++ pyanaconda/modules/payloads/packages/packages.py
@@ -22,7 +22,7 @@ from pyanaconda.core.signal import Signal
 from pyanaconda.core.configuration.anaconda import conf
 from pyanaconda.modules.common.base import KickstartBaseModule
 from pyanaconda.modules.common.constants.objects import PAYLOAD_PACKAGES
-from pyanaconda.modules.common.errors import InvalidValueError, UnsupportedValueError
+from pyanaconda.modules.common.errors.general import InvalidValueError, UnsupportedValueError
 from pyanaconda.modules.payloads.packages.constants import MultilibPolicy, TIMEOUT_UNSET, \
     RETRIES_UNSET, LANGUAGES_DEFAULT, LANGUAGES_NONE
 from pyanaconda.modules.payloads.packages.packages_interface import PackagesInterface
