+++ tests/nosetests/pyanaconda_tests/module_payload_packages_test.py
@@ -29,7 +29,7 @@ from dasbus.typing import *  # pylint: disable=wildcard-import
 
 from pyanaconda.core.configuration.payload import PayloadSection
 from pyanaconda.modules.common.constants.objects import PAYLOAD_PACKAGES
-from pyanaconda.modules.common.errors import InvalidValueError, UnsupportedValueError
+from pyanaconda.modules.common.errors.general import InvalidValueError, UnsupportedValueError
 from pyanaconda.modules.payloads.payloads import PayloadsService
 from pyanaconda.modules.payloads.payloads_interface import PayloadsInterface
 from pyanaconda.modules.payloads.packages.packages import PackagesModule
